capitulo-03-tratamentos-de-erros-e-testes-automatizados/aula-03.022-implementando-endpoint-da-edicao/beerstore/src/main/java/com/hibicode/beerstore/service/BeerService.java
package com.hibicode.beerstore.service;

import com.hibicode.beerstore.model.Beer;
import com.hibicode.beerstore.repository.Beers;
import com.hibicode.beerstore.service.exception.BeerAlreadyExistException;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import javax.persistence.EntityNotFoundException;
import java.util.Optional;


@Service
public class BeerService {

//    public void save(Beer beer) {
//        throw new BeerAlreadyExistException();
//    }

//    @Autowired
    private Beers beers;

    public BeerService(@Autowired Beers beers) {
        this.beers = beers;
    }

    public Beer save(final Beer beer) {
        this.verifyIfBeerExists(beer);
        return this.beers.save(beer);
    }

    public void delete(final Long id) {
        final Optional<Beer> beerById = this.beers.findById(id);
        if (!beerById.isPresent()) {
            throw new EntityNotFoundException();
        }
        this.beers.delete(beerById.get());
    }

    private void verifyIfBeerExists(final Beer beer) {
        Optional<Beer> beerByNameAndType = this.beers.findByNameAndType(beer.getName(), beer.getType());

        if (beerByNameAndType.isPresent() && (beer.isNew() || isUpdatingToDifferentBeer(beer, beerByNameAndType))) {
            throw new BeerAlreadyExistException();
        }
    }

    private boolean isUpdatingToDifferentBeer(Beer beer, Optional<Beer> beerByNameAndType) {
        return beer.alreadyExist() && !beerByNameAndType.get().equals(beer);
    }
}
