capitulo-03-tratamentos-de-erros-e-testes-automatizados/aula-03.016-criando-teste-para-salvar-cerveja-com-hamcrest/beerstore/src/test/java/com/hibicode/beerstore/service/BeerServiceTest.java
package com.hibicode.beerstore.service;

import com.hibicode.beerstore.model.Beer;
import com.hibicode.beerstore.model.BeerType;
import com.hibicode.beerstore.service.exception.BeerAlreadyExistException;
import org.junit.Before;
import org.junit.Test;

import java.math.BigDecimal;

import static org.hamcrest.MatcherAssert.assertThat;

import static org.hamcrest.Matchers.*;

public class BeerServiceTest {

    private BeerService beerService;

    @Before
    public void setUp() {
        this.beerService = new BeerService();
    }

    @Test(expected = BeerAlreadyExistException.class)
    public void should_deny_creatin_of_beer_that_exists() {
//        BeerService beerService = new BeerService();

        Beer newBeer = new Beer();
        newBeer.setName("Heineken");
        newBeer.setType(BeerType.LAGER);
        newBeer.setVolume(new BigDecimal("355"));

        beerService.save(newBeer);
    }

    @Test
    public void should_create_new_beer() {
//        BeerService beerService = new BeerService();

        Beer newBeer = new Beer();
        newBeer.setName("Heineken");
        newBeer.setType(BeerType.LAGER);
        newBeer.setVolume(new BigDecimal("355"));

        Beer beerSaved = beerService.save(newBeer);

        assertThat(beerSaved.getId(), equalTo(10L));
        assertThat(beerSaved.getName(), equalTo("Heineken"));
        assertThat(beerSaved.getType(), equalTo(BeerType.LAGER));
    }
}
